requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-jalali-date==1.0.1
jdatetime==4.1.0
Pillow==9.2.0
sqlparse==0.4.2
